# internal/apidocs/swagger.yaml
basePath: /api/v1
definitions:
  app.HealthData:
    properties:
      components:
        additionalProperties:
          type: string
        type: object
      status:
        type: string
    type: object
  auth.AuthStatusResponse:
    properties:
      initialized:
        type: boolean
    type: object
  auth.InitAdminRequest:
    properties:
      confirmPassword:
        type: string
      password:
        maxLength: 64
        minLength: 8
        type: string
      username:
        maxLength: 32
        minLength: 3
        type: string
    required:
    - confirmPassword
    - password
    - username
    type: object
  auth.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  auth.UserResponse:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      username:
        type: string
    type: object
  camera.CameraCredentialsResponse:
    properties:
      cameraId:
        type: string
      credentials:
        additionalProperties:
          type: string
        description: role -> full plaintext RTSP URL
        type: object
    type: object
  camera.CameraResponse:
    properties:
      createdAt:
        type: string
      degraded:
        type: boolean
      enabled:
        type: boolean
      health:
        type: string
      id:
        type: string
      lastCheckedAt:
        type: string
      lastSuccessAt:
        type: string
      name:
        type: string
      reason:
        type: string
      revision:
        type: integer
      session:
        type: string
      stale:
        type: boolean
      streams:
        items:
          $ref: '#/definitions/camera.StreamResponse'
        type: array
      updatedAt:
        type: string
    type: object
  camera.CameraStateInfo:
    properties:
      cameraId:
        type: string
      degraded:
        type: boolean
      enabled:
        type: boolean
      health:
        type: string
      lastCheckedAt:
        type: string
      lastSuccessAt:
        type: string
      reason:
        type: string
      revision:
        type: integer
      session:
        type: string
      stale:
        type: boolean
      streams:
        additionalProperties:
          $ref: '#/definitions/camera.StreamStateInfo'
        type: object
    type: object
  camera.CreateCameraRequest:
    properties:
      enabled:
        type: boolean
      id:
        type: string
      mainStream:
        $ref: '#/definitions/camera.CreateStreamRequest'
      name:
        maxLength: 64
        minLength: 1
        type: string
      subStream:
        $ref: '#/definitions/camera.CreateStreamRequest'
    required:
    - mainStream
    - name
    type: object
  camera.CreateStreamRequest:
    properties:
      protocol:
        description: '"rtsp" (default)'
        type: string
      role:
        description: '"main" or "sub"'
        type: string
      rtspUrl:
        type: string
      transport:
        description: '"tcp" or "udp" (default "tcp")'
        type: string
    required:
    - rtspUrl
    type: object
  camera.DiagnoseResponse:
    properties:
      cameraId:
        type: string
      message:
        type: string
      state:
        $ref: '#/definitions/camera.CameraStateInfo'
    type: object
  camera.StreamResponse:
    properties:
      codec:
        type: string
      createdAt:
        type: string
      fps:
        type: number
      fpsString:
        type: string
      height:
        type: integer
      id:
        type: integer
      protocol:
        type: string
      role:
        type: string
      rtspUrl:
        description: Complete RTSP URL for frontend playback and display
        type: string
      runtimeState:
        $ref: '#/definitions/camera.StreamStateInfo'
      transport:
        type: string
      updatedAt:
        type: string
      width:
        type: integer
    type: object
  camera.StreamStateInfo:
    properties:
      consecutiveFailures:
        type: integer
      errorMessage:
        type: string
      evidenceType:
        type: string
      health:
        type: string
      lastCheckedAt:
        type: string
      lastSuccessAt:
        type: string
      reason:
        type: string
      role:
        type: string
      session:
        type: string
    type: object
  camera.UpdateCameraRequest:
    properties:
      enabled:
        type: boolean
      mainStream:
        $ref: '#/definitions/camera.CreateStreamRequest'
      name:
        maxLength: 64
        minLength: 1
        type: string
      revision:
        type: integer
      subStream:
        $ref: '#/definitions/camera.CreateStreamRequest'
    required:
    - revision
    type: object
  httputil.FieldDetail:
    properties:
      code:
        type: string
      field:
        type: string
      message:
        type: string
    required:
    - code
    - field
    - message
    type: object
  httputil.PaginatedData-camera_CameraResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/camera.CameraResponse'
        type: array
      limit:
        description: Backwards compatibility fields for legacy clients expecting limit/offset
        type: integer
      offset:
        type: integer
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
  httputil.Response:
    properties:
      code:
        type: string
      data: {}
      details:
        items:
          $ref: '#/definitions/httputil.FieldDetail'
        type: array
      message:
        type: string
    required:
    - code
    - data
    - message
    type: object
info:
  contact: {}
  description: Local API for the Zhulong edge operations host.
  title: Zhulong API
  version: "1.0"
paths:
  /auth/init:
    post:
      consumes:
      - application/json
      parameters:
      - description: Admin credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.InitAdminRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httputil.Response'
            - properties:
                data:
                  $ref: '#/definitions/auth.UserResponse'
              type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httputil.Response'
        "413":
          description: Request Entity Too Large
          schema:
            $ref: '#/definitions/httputil.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/httputil.Response'
      summary: Initialize administrator account
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httputil.Response'
            - properties:
                data:
                  $ref: '#/definitions/auth.UserResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.Response'
        "413":
          description: Request Entity Too Large
          schema:
            $ref: '#/definitions/httputil.Response'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/httputil.Response'
      summary: Login as administrator
      tags:
      - auth
  /auth/logout:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httputil.Response'
            - properties:
                data:
                  type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.Response'
      summary: Logout administrator session
      tags:
      - auth
  /auth/me:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httputil.Response'
            - properties:
                data:
                  $ref: '#/definitions/auth.UserResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.Response'
      summary: Get currently authenticated user
      tags:
      - auth
  /auth/status:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httputil.Response'
            - properties:
                data:
                  $ref: '#/definitions/auth.AuthStatusResponse'
              type: object
      summary: Get system initialization status
      tags:
      - auth
  /cameras:
    get:
      parameters:
      - description: Page number (1-based, default 1)
        in: query
        name: page
        type: integer
      - description: Page size (default 20, max 100)
        in: query
        name: pageSize
        type: integer
      - description: Legacy pagination limit
        in: query
        name: limit
        type: integer
      - description: Legacy pagination offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httputil.Response'
            - properties:
                data:
                  $ref: '#/definitions/httputil.PaginatedData-camera_CameraResponse'
              type: object
      summary: List cameras
      tags:
      - camera
    post:
      consumes:
      - application/json
      parameters:
      - description: Camera parameters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/camera.CreateCameraRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/httputil.Response'
            - properties:
                data:
                  $ref: '#/definitions/camera.CameraResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/httputil.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/httputil.Response'
      summary: Create a new camera
      tags:
      - camera
  /cameras/{id}:
    delete:
      parameters:
      - description: Camera ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httputil.Response'
            - properties:
                data:
                  type: object
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.Response'
      summary: Delete camera
      tags:
      - camera
    get:
      parameters:
      - description: Camera ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httputil.Response'
            - properties:
                data:
                  $ref: '#/definitions/camera.CameraResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.Response'
      summary: Get camera by ID
      tags:
      - camera
    put:
      consumes:
      - application/json
      parameters:
      - description: Camera ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated parameters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/camera.UpdateCameraRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httputil.Response'
            - properties:
                data:
                  $ref: '#/definitions/camera.CameraResponse'
              type: object
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/httputil.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/httputil.Response'
      summary: Update camera configuration
      tags:
      - camera
  /cameras/{id}/credentials:
    get:
      parameters:
      - description: Camera ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httputil.Response'
            - properties:
                data:
                  $ref: '#/definitions/camera.CameraCredentialsResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.Response'
      summary: Get camera plaintext credentials
      tags:
      - camera
  /cameras/{id}/diagnose:
    post:
      parameters:
      - description: Camera ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httputil.Response'
            - properties:
                data:
                  $ref: '#/definitions/camera.DiagnoseResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.Response'
      summary: Manually diagnose camera connection
      tags:
      - camera
  /cameras/{id}/streams/{role}/ws:
    get:
      parameters:
      - description: 摄像机 ID
        in: path
        name: id
        required: true
        type: string
      - description: 码流类型 (main 或 sub，默认 main)
        in: path
        name: role
        type: string
      responses: {}
      summary: WebSocket 视频流实时预览
      tags:
      - camera
  /cameras/events:
    get:
      produces:
      - text/event-stream
      responses:
        "200":
          description: Server-Sent Events stream
          schema:
            type: string
      summary: Subscribe to camera state events via SSE
      tags:
      - camera
  /health:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httputil.Response'
            - properties:
                data:
                  $ref: '#/definitions/app.HealthData'
              type: object
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/httputil.Response'
      summary: Get host health
      tags:
      - system
swagger: "2.0"